+++ CHANGELOG.md
@@ -11,10 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
- - [Core] Add `--i18n-keywords` and `--i18n-languages` options ([#3053](https://github.com/cucumber/cucumber-jvm/pull/3053) M.P. Korstanje)
+- [Core] Add `--i18n-keywords` and `--i18n-languages` options ([#3053](https://github.com/cucumber/cucumber-jvm/pull/3053) M.P. Korstanje)
+- [JUnit Platform Engine] Warn when selected line does not exist ([#3056](https://github.com/cucumber/cucumber-jvm/pull/3056) M.P. Korstanje)
 
 ### Deprecated
- - [Core] Deprecate `--i18n` options ([#3053](https://github.com/cucumber/cucumber-jvm/pull/3053) M.P. Korstanje)
+- [Core] Deprecate `--i18n` options ([#3053](https://github.com/cucumber/cucumber-jvm/pull/3053) M.P. Korstanje)
 
 ## [7.27.2] - 2025-08-18
 ### Fixed