+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [Core] Fix NPE for optional arguments when using the Cucumber JSON Formatter ([cucumber-json-formatter/#7](https://github.com/cucumber/cucumber-json-formatter/pull/7), [#3060](https://github.com/cucumber/cucumber-jvm/pull/3060) M.P. Korstanje)
 
 ## [7.28.0] - 2025-09-01
 ### Added
@@ -19,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [JUnit Platform Engine] Support rerun files ([#3057](https://github.com/cucumber/cucumber-jvm/pull/3057) M.P. Korstanje)
 
 ### Changed
-- [Core] Use a [message based Cucumber JSON Formatter](https://github.com/cucumber/cucumber-json-formatter) ([##2888](https://github.com/cucumber/cucumber-jvm/pull/#2888) M.P. Korstanje)
+- [Core] Use a [message based Cucumber JSON Formatter](https://github.com/cucumber/cucumber-json-formatter) ([#2888](https://github.com/cucumber/cucumber-jvm/pull/#2888) M.P. Korstanje)
 
 ### Deprecated
 - [Core] Deprecate `--i18n` options ([#3053](https://github.com/cucumber/cucumber-jvm/pull/3053) M.P. Korstanje)